requirements.txt
streamlit==1.29.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.3
markdown==3.5.1
requests==2.31.0
